Battle effect rendering: flame-atlas upload, sprite-atlas field-order fix, summon-animation correction (#145)
* Battle effect VRAM: pin + upload the PROT 870 flame-texture atlas

PROT 870 (CDNAME-mislabeled `sound_data`) is three back-to-back 64x256
4bpp PSX TIMs that target VRAM (320,0)/(384,0)/(448,0) with CLUTs in
rows 474..476 - the battle fire/flame effect-texture atlas. Byte-verified
pixel-exact against every stable Rim Elm battle capture (command-menu,
submenu, pre/post Seru-capture frames all match 100%; a still-loading
frame matches only the mid-DMA snapshot). Distinct from etim.dat (PROT
874 section 2, pages at fb_y=256): PROT 870 sits at fb_y=0 in the same
VRAM columns the field uses for town stage textures, so it is a
battle-only upload (the town01 field captures hold unrelated texels
there) and is not pulled by the FUN_800520F0 etmd/befect load path.

- engine-core: `scene::upload_flame_atlas_into_vram` walks the three
  TIMs out of the entry's extended footprint and blits them (image +
  optional CLUT). Disc-gated test pins exactly 3 TIMs at the verified
  targets.
- play-window: blit the atlas on battle entry into the throwaway VRAM
  copy that battle exit discards, so the resident field VRAM is never
  clobbered (mirrors retail overwriting that region for battle).
- docs + site: document PROT 870's layout/target/load characteristics
  in effect.md / effect-vm.md / open-rev-eng-threads, and record the
  finding that the separate efect.dat 2D-billboard page-(0,0) 8bpp
  source coincides with the live battle framebuffer - so resolving it
  needs a capture of a 2D-billboard effect firing, which the savestate
  corpus lacks (capture-blocked, not resolvable from current captures).


* Effect billboards: resolve the texel source (atlas tpage is relocated, not page (0,0))

A melee hit-spark battle capture resolves the long-open "what fills the
efect.dat sprite-atlas page (0,0)" question. Walking the live GPU prim
pool on the impact frame: no on-screen primitive samples page (0,0),
8bpp, or tpage 0x7680 anywhere (there are zero 8bpp textured prims in
the whole battle). The atlas's stored tpage 0x7680 is an authoring
placeholder the runtime relocates at spawn.

The white hit-spark draws as textured quads (POLY_FT4/GT4) sampling the
loaded effect-texture pages - specifically the PROT 870 flame-atlas
pages (320,0)/(448,0) with effect-band CLUTs (rows 473..480). Those
pages appear only in the impact frame; a no-spark command-menu frame
from the same fight has only the persistent party/monster mesh pages.
So the billboard texel source is the same PROT 870 + etim effect
textures the loader populates; page (0,0) is just the framebuffer. The
remaining (narrowed) work is the per-spawn rule that rewrites 0x7680 to
the real effect page, so the engine can sample it instead of the literal
placeholder.

Also catalogues three new battle-effect captures in scripts/scenarios.toml
(by content fingerprint; the save binaries stay out of the repo): the
melee hit-spark frame and a Gimard Tail Fire frame pair. The Tail Fire
pair's effect-CLUT band is byte-identical between frames, pointing the
flame flicker at per-frame primitive CBA-column cycling rather than
CLUT-data re-upload.

Docs + site (effect.md / effect-vm.md / open-rev-eng-threads) updated to
the resolved finding.


* Effect billboards: fix the sprite-atlas CLUT/tpage field order

The effect sprite-atlas entry was parsed with its CLUT and tpage fields
swapped. The pass-2 emit (FUN_801E0088, ~0x801E0980) writes the u16 at
atlas+4 into the GPU primitive's CLUT field (POLY_FT4 word3 high) and
the byte at atlas+6 into its tpage field (word5 high) - so the entry is
[u8 u, v, w, h, u16 CLUT@+4, u8 tpage@+6, u8 unk], the reverse of what
the struct and docs claimed.

This is what made the billboards look like they sampled VRAM page (0,0),
8bpp: 0x7680 was being read as a tpage, but it is the CLUT - as a CBA it
decodes to (0, 474), an effect-CLUT row (PROT 870 TIM0's CLUT). The real
tpage is the +6 byte (e.g. 0x25 = page (320,0), 4bpp). A melee hit-spark
battle capture confirms it: no primitive samples page (0,0)/8bpp, and the
spark draws as textured quads on the PROT 870 flame-atlas pages
(320,0)/(448,0) with effect-band CLUTs.

SpriteAtlasEntry now reads clut (u16) from +4 and the tpage byte from +6;
the EffectSprite push drops the now-redundant cast; the play-window
cba_tsb pair was already [clut, page] so it becomes correct once the
fields hold the right values. With PROT 870 + etim resident in battle
VRAM, World::active_effect_sprites now yields the real effect page + CLUT
and the billboards sample resident texels instead of empty page (0,0).

Tests: a real-data field-order regression in effect_catalog_real.rs
asserts the atlas CLUTs land in the effect-CLUT band and the tpages pick
effect pages (fails if the fields are ever swapped back); synthetic atlas
tests updated. Docs + site corrected.


* Effect summon: correct the flame animation mechanism (geometric, not CLUT cycling)

Decompiling the Gimard summon overlay (PROT 905, FUN_801F69D8 at base
0x801F69D8) and diffing two animation-distinct Tail Fire capture frames
falsifies the "flame flicker = per-frame CLUT/palette cycling" reading.

The two frames are clearly different animation moments (framebuffer
pixels differ ~21%) yet their entire CLUT band (VRAM rows 470..499) is
byte-identical - so no per-frame palette or CBA cycling occurs at all.
The visible flame motion is geometric: the overlay is a move-VM
scene-graph that stages ~22 part-actors (FUN_80021B04) and animates
them via FUN_80023070 + 0x801D829C tweens.

The overlay does contain 3 conditional LoadImage (0x800583C8) CLUT
uploads, but they build RECT {x=0, y=481+s5, w=240, h=1} from source
a2 + s5*480 + 0x894 - targeting VRAM row 481+ (the character/party-CLUT
region), not the flame's row 478, and that region is also byte-identical
across the two frames. So they are a conditional secondary effect, not
the flame animation.

Consequence: there is no faithful CLUT/CBA cycle to wire (doing so would
fabricate an effect retail does not perform). The engine already renders
the static flame mesh with the correct baked row-478 CLUT; faithfully
animating it requires porting the PROT 905 move-VM summon scene-graph
(geometric), which is deferred as a larger effort.

Docs, site, and the open-thread index corrected; no code change.


---------

+++ crates/engine-core/src/world.rs
@@ -3179,7 +3179,7 @@ impl World {
                     uv: [e.u as u16, e.v as u16],
                     uv_size: [e.w as u16, e.h as u16],
                     page: e.page,
-                    clut: e.clut as u16,
+                    clut: e.clut,
                     age01,
                 });
             }
